Display epoch progression during training

+++ src/utils/printing.h
@@ -8,8 +8,8 @@
 #include <string>
 
 // Construct progress bar of the form [==============>-----]
-std::string constructProgressBar(int currentStep, int maxSteps);
+std::string constructProgressBar(size_t currentStep, size_t maxSteps);
 // Construct percentage in the form [25/100 (25%)]
-std::string constructPercentage(int currentStep, int maxSteps);
+std::string constructPercentage(size_t currentStep, size_t maxSteps);
 
 #endif //NNLIB_PRINTING_H
